services/collector/requirements.txt
kafka-python==2.0.2
pyyaml==6.0.2
requests==2.31.0
websocket-client==1.7.0
betterproto==2.0.0b6
mini-racer==0.12.4
